refactor: Collector Moudile refactor
