+++ collector/src/main/java/com/whalewatch/service/WebsocketService.java
@@ -1,6 +1,6 @@
 package com.whalewatch.service;
 
-import com.whalewatch.ExchangesProperties;
+import com.whalewatch.service.config.ExchangesProperties;
 import org.springframework.web.socket.WebSocketHandler;
 
 import java.util.Map;
