+++ collector/src/main/java/com/whalewatch/service/dto/TradeDto.java
@@ -1,4 +1,4 @@
-package com.whalewatch;
+package com.whalewatch.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
